+++ app/controllers/admin/users_controller.rb
@@ -1,8 +1,4 @@
-class Admin::UsersController < ApplicationController
-  layout 'admin'
-
-  before_action :authenticate_user!
-  before_action :admin_required
+class Admin::UsersController < AdminController
 
   def index
     @users = User.all
