+++ app/controllers/admin_controller.rb
@@ -1,0 +1,6 @@
+class AdminController < ApplicationController
+  layout 'admin'
+
+  before_action :authenticate_user!
+  before_action :admin_required
+end
